btn-click-updated
- used modules for the constants and function.
- created one common module for all the constants and functions use.
- padding was used with short-hand
