+++ btnclick.htm
@@ -10,7 +10,6 @@
 
     <!--linking js file via defer-->
     <script src="js/btnclick.js" type="module"></script>
-    <script src="js/constants.js" type="module"></script>
   </head>
