+++ js/domconstants.js
@@ -1,0 +1,6 @@
+let shapes = document.querySelector(".shapes");
+let next_btn = document.querySelector(".nxt_btn");
+let header = document.querySelector(".learning_header");
+let mainContent = document.querySelector(".learning_content");
+
+export { shapes, next_btn, header, mainContent };
