+++ styles/css/btn-click.css
@@ -175,8 +175,7 @@ body {
   width: 130px;
   height: 50px;
   box-sizing: border-box;
-  padding-top: 15px;
-  padding-left: 15px;
+  padding: 15px 0 0 15px;
 }
 
 /* first column of the table structure */
@@ -307,10 +306,7 @@ body {
   border: 2px solid #ffffff;
   background-color: #000000;
   color: #ffffff;
-  padding-left: 30px;
-  padding-right: 30px;
-  padding-top: 10px;
-  padding-bottom: 10px;
+  padding: 10px 30px 10px 30px;
   font-weight: 900;
   font-size: 20px;
   cursor: pointer;
